# Optimize and fix issues with the `Parser` module