--- a/data/jsonchecker/fail32.json
+++ b/data/jsonchecker/fail32.json
@@ -1,1 +1,1 @@
-{"Comma instead if closing brace": true,
+{"Comma instead of closing brace": true,