--- a/data/jsonchecker/fail34.json
+++ b/data/jsonchecker/fail34.json
@@ -1,0 +1,1 @@
+{"garbage" before : "separator"}