--- a/data/jsonchecker/fail35.json
+++ b/data/jsonchecker/fail35.json
@@ -1,0 +1,1 @@
+{"no separator"