--- a/data/jsonchecker/fail36.json
+++ b/data/jsonchecker/fail36.json
@@ -1,0 +1,1 @@
+{"no closing brace": true