Overview of changes:

 - Force parse_string to return Compat.UTF8String
 - Don't use IOBuffer objects in parse_string
 - Sprinkled @inbounds and @inline liberally
 - Make default Dict key typed Compat.UTF8String instead of Any
 - Fix some unicode compatibility along the way
 - Folded the streaming parser into the regular parser, using generic implementations (optimizations for number and string parsing in-memory)

Impact:

Speed for parsing both strings and non-strings have improved substantially on my computer. By my tests we are now easily within a factor of 2 from Python's json with or without strings using [@samuelcolvin](https://gist.github.com/samuelcolvin/11101522)'s benchmarks. Note that these benchmarks heavily rely on float parsing which is still slow, and are ASCII-only which obscures the biggest performance gain obtained.

Here are the new benchmark results. Note that these benchmarks have been "over-benchmarked" to some degree, since they were used heavily in deciding which optimizations to pursue. However, they should in theory represent the kinds of data one would parse quite well. Python's builtin json is somewhat unfairly penalized on these benchmarks, because it is pitifully slow on unicode data. ujson on the other hand is quite good at it. However, ujson "cheats" by default on the numeric benchmarks by not parsing floats to best precision; I included the "non-cheating" version too for completeness.

```
Julia 0.5:
 [Bench] canada 0.057967427 seconds
 [Bench] citm_catalog 0.027227154 seconds
 [Bench] citylots 6.882568502 seconds
 [Bench] twitter 0.011117001 seconds
 [Bench] Total (G.M.) 0.10482892077727128 seconds

Python 2 json:
python json canada time: 0.050s
python json citm_catalog time: 0.032s
python json citylots time: 8.600s
python json twitter time: 0.552s
GM 0.295 (2.81 times Julia)

Python 2 ujson:
python ujson canada time: 0.026s
python ujson citm_catalog time: 0.014s
python ujson citylots time: 5.535s
python ujson twitter time: 0.005s
GM 0.056 (0.534 times Julia)

Python 2 ujson (no cheating):
python ujson canada time: 0.045s
python ujson citm_catalog time: 0.014s
python ujson citylots time: 6.687s
python ujson twitter time: 0.005s
GM 0.068 (0.649 times Julia)

Old Julia 0.5
 [Bench] canada 0.100768632 seconds
 [Bench] citm_catalog 0.089031051 seconds
 [Bench] citylots 16.282590906 seconds
 [Bench] twitter 0.043230788 seconds
 [Bench] Total (G.M.) 0.2819005229832309 seconds (2.69 times new Julia)
```

This change involves a rewrite of object, number, and array parsing code, which also makes things more correct. It removes all of the allowed failures for JSON checker.

Breakage:

The streaming parsers no longer keep track of what has been parsed, and so the error messages are now effectively worthless.

Closes #98, #118, #122, #127.