@TotalVerb 2016-05-01
A lot of the remaining time for the no-strings test is due to the type instability of `parse_number`. I am wondering whether it is worth it to simply return `Float64` always from this method. JSON itself does not distinguish between integers and decimals, and JavaScript uses `Float64` for everything.

I will try out some speculative typing for arrays, which might still give a small performance boost.