@kmsquire 2016-05-01
@TotalVerb, thanks for doing this!  I'll note that this package sometimes doesn't get much attention, so if you find this PR languishing, please bump it once or twice (or more).