@TotalVerb 2016-05-09
This will need an update to support the latest master's String changes. I plan to work on that later tonight.