@stevengj 2016-05-15
cc @StefanKarpinski, who will be interested to hear of the `String` change impacts.