@TotalVerb 2016-05-16
Actually, I think I was mistaken about the source of the performance regression. `nextind` probably isn't at fault.