@TotalVerb 2016-05-16
Here is a condensed benchmark that suffers from significant regression:

```
function f(s)
    n = 0
    m = start(s)
    e = endof(s)
    while m ≤ e
        n += Int(s[m])
        m = nextind(s, m)
    end
    n
end
```

Pre-stringapalooza, on `UTF8String` this is over 10 times as slow as on `ASCIIString`. `String` seems to have inherited `UTF8String`'s speed.