@yuyichao 2016-05-16
It's not surprising that the new `String` inherits `UTF8String`'s speed since that's exactly what it is.

I've also seen similar effect in https://groups.google.com/forum/?fromgroups=#!topic/julia-users/9zUDciJk878 where using `nextind` shows a 2x slow down. I remember @StefanKarpinski was talking about more efficient iteration on string index and I was expecting it to come in one of the future related changes.