@TotalVerb 2016-05-20
I think the major issues (excepting compatibility with v0.3) are sorted out now. I'll see if reusing buffers still makes sense, and then I'll make sure all the code additions are actually having substantial impact.