@TotalVerb 2016-05-22
Non-test code has grown by about 20 to 30 lines. (Some of the added lines are comments.) However, given that a very good chunk of the tests that failed before now pass, I think the slight increase in LOC is reasonable. The only functions which have seen their complexity increase are `parse_string` and `parse_simple`. I don't know if there's a way to keep `parse_string` as simple as it was before while maintaining correctness and efficiency, at least until `nextind` performance becomes much better. `parse_simple` on the other hand was behaving quite oddly earlier (accepting `tXXe` as true, which for some reason was even tested?!), and I think it is better to stick to the JSON standard at some small complexity cost.

The only downside to this change (as far as I can see) is that `parse_number` has less range on 64-bit machines, since it now always returns a `Float64`. (Before, it could parse 64-bit integers on 64-bit machines... but would also crash on those same integers on 32-bit machines, so it is debatable whether that behaviour was good.) It might be useful, in a later PR, to let the user supply a custom parse_number function. Especially now that functions can be specialized on, this would be similar to how a dict type can be passed in currently.

v0.3 compatibility is back, but the timing is slightly worse than on newer versions. I don't think that is a big deal.