@TotalVerb 2016-05-22
I was wondering why the array and object parsing code seems slower recently. It turns out that a recent refactor (moving the burden of throwing the error to a helper function) has led to the construction of a bunch of temporary strings. This can be fixed by removing the second argument, but unfortunately the tailored error messages (like `Expected a '{' here`) will be gone too. A macro would work too, and preserve the tailored error messages, but that looks ugly for some reason...