@tkelman 2016-05-22
could delay constructing the string until the error gets displayed?