@TotalVerb 2016-05-24
Should `tryparse` maybe also go in Compat.jl? Then we could be rid of almost all compatibility code.