@TotalVerb 2016-05-24
I think the `Int`/`Float64` change may be more breaking than I had expected. At least [Transit.jl](https://github.com/russolsen/Transit.jl) seems to rely on the old behaviour.

It's probably best to revert the type-stability change for now, until there's actually a way to specify the number parse function. (However, a good change would be to swap `Int64` for `Int`, so that 32-bit builds do not crash.)