@kmsquire 2016-05-24
> I think the `Int/Float64` change may be more breaking than I had expected. At least `Transit.jl` and any package expecting Ints to roundtrip, may be relying on that behaviour.

Since it's not in the JSON standard, packages shouldn't really be relying on it.  A deprecation warning would be appropriate.

(To be fair, I argued differently when originally introducing `OrderedDict`'s...)