@TotalVerb 2016-05-25
There is no real way to deprecate producing `Int`s as numbers without deprecating a whole lot of valid JSON. I think I'll run the tests for some packages that depend on JSON.jl, and see if too much breaks.