@kmsquire 2016-05-26
Not that we have to follow Python, but I wasn't aware that they converted to int.  I'm more ok with us doing so as well then.