@TotalVerb 2016-05-26
I can't reproduce the 0.3 Travis failure :confused:.