@TotalVerb 2016-05-27
I restored the int-parsing behaviour. Performance has degraded a little bit, as expected. But overall it's not so bad, and for some reason it looks like compilation time is much lower. We also can get a bit of performance gain from using `DataStructures.OrderedDict`, if users are inclined to squeeze out every last drop of performance.