@TotalVerb 2016-05-28
Great news! `nextind` is quite costly in many parts of the code, and optimizing that has improved the geometric mean benchmark results on my machine to *0.17*. This is even lower than before parse_number type stability was reverted.

This particular change might be applicable to Base too, as it's an almost 3-fold speed improvement over the current implementation of `nextind`.