@stevengj 2016-05-28
In this particular application, however, it's not clear to me why you need `nextind` at all.   Since parsing JSON only involves looking for ASCII characters and matching substrings, you should be able to do nearly everything by operating on the raw `s.data` byte array, rather than working with Unicode chars.