@TotalVerb 2016-05-28
@stevengj Yes, you're right. In fact some of the performance gain comes from avoiding nextind in string and number parsing. Unfortunately these improved versions somewhat duplicate the generic versions, which provide a performance and memory advantage when parsing a stream.

I did consider avoiding it everywhere, but that currently makes code somewhat messy (== @compat UInt('x') everywhere), and would require a rethink of the streaming parser.

But to get more performance, that seems to be necessary. I'll see if I can figure something out.