@stevengj 2016-05-30
(Can some of the byte-based parsing optimizations be folded back in to Base?  e.g. number parsing of UTF-8 strings seems like it could always avoid `nextind` as long as only ASCII digits are allowed.)