@stevengj 2016-05-30
I guess you could define `const _x = @compat UInt('x')` to simplify the code somewhat.