@TotalVerb 2016-05-30
Base already has a fast strtod, which we're using here. No idea about the other stuff.

The consts are a good idea. I'll play around until I have a solution.