@TotalVerb 2016-05-31
I got a huge performance bonus by not copying the buffer data when creating `String`s. Unfortunately, the way I did this was somewhat sketchy:

```
ps.s = s
ret = Compat.UTF8String(UInt8[])
b = sizehint!(ret.data, len)
parse_string(ps, b)
ret
```

Is there a better way to use a known-good UTF8 buffer to create a `String` without making a copy?